requirements.txt
black
ruff
python-dotenv
python-telegram-bot
